+++ index.js
@@ -71,6 +71,7 @@ function initCrossword(xmlData) {
   }
 
   document.addEventListener('keydown', (e) => crossword.handleKeyDown(e));
+  document.addEventListener('beforeinput', (e) => crossword.handleBeforeInput(e));
   document.addEventListener('input', (e) => crossword.handleInput(e));
 
   crossword.buildGrid();